// apps/backend/src/enums/subjects.enum.ts
export enum Subjects {
  USER = 'User',
  USER_ADMIN = 'UserAdmin',
  USER_ROLE = 'UserRole',

  FILE = 'File',
  FILE_ADMIN = 'FileAdmin',

  NOTIFICATION = 'Notification',

  ADMIN_DASHBOARD = 'AdminDashboard',
  ADMIN_ACCESS_LOG = 'AdminAccessLog',
  ADMIN_SETTINGS = 'AdminSettings',
  CAPTCHA_ADMIN = 'CaptchaAdmin',
}